+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.1
+
+### Bug Fixes
+
+* Fixing duplicate Key creation
+* Using unique key alias (per package name)
+* Expose KeyStore helper method (Unlock)
+
+### Enhancements
+
+* Add more tests coverage
+
 ## 1.0.0
 
-Initial release.
+* Initial release.